requirements.txt
annotated-types==0.7.0
anyio==4.10.0
base58==2.1.1
beanie==2.0.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
coincurve==21.0.0
cytoolz==1.0.1
dnspython==2.7.0
eth-hash==0.7.1
eth-typing==5.2.1
eth-utils==5.3.1
eth_abi==5.2.0
exceptiongroup==1.3.0
fastapi==0.116.1
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
lazy-model==0.3.0
motor==3.7.1
parsimonious==0.10.0
pycryptodome==3.23.0
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pymongo==4.14.1
python-dotenv==1.1.1
python-telegram-bot==22.3
PyYAML==6.0.2
regex==2025.8.29
requests==2.32.5
sniffio==1.3.1
starlette==0.47.3
toolz==1.0.0
tronpy==0.6.1
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.35.0
# uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
